flask-service/requirements.txt
Flask
Flask-Cors
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
gunicorn
numpy
pandas
requests
Pillow
scikit-learn==1.3.2
tensorflow
tf-keras
torch
transformers
opencv-python
ultralytics

